services/osint/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
bbot==2.1.7
neo4j==5.23.1
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
orjson==3.10.7
ujson==5.10.0
loguru==0.7.2
pyyaml==6.0.2
typing_extensions==4.12.2

# MCP server library (Model Context Protocol)
mcp==1.1.0

